+++ src/main/java/grupo/quatro/api_manage_escola/Receive/Boletim/DadosRegistrarBoletim.java
@@ -20,6 +20,6 @@ public record DadosRegistrarBoletim(
         @NotNull
         Bimestre bimestre,
 
-        Long nota
+        Double nota
 ) {
 }
